chore: organize test dir
